refactor: accounts routes
