+++ .vscode/settings.json
@@ -7,6 +7,7 @@
     "delegators",
     "fastify",
     "milkomeda",
+    "mirs",
     "nutcoin",
     "nutlink",
     "openapi",
